[REFACTOR] search page, detail page with redux interviewees state using byId
