+++ src/redux/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 import { all } from "redux-saga/effects";
 
-import interviewees, { intervieweeSaga } from "./interviewee";
+import interviewees, { intervieweeSaga } from "./interviewees";
 import projects, { projectsSaga } from "./projects";
 import user, { userSaga } from "./user";
